added error checking
